Extract interfaces to be implemented by RPC as well.
