+++ src/main/java/no/fint/p360/data/p360/soap/P360AbstractSOAPService.java
@@ -1,4 +1,4 @@
-package no.fint.p360.data.p360;
+package no.fint.p360.data.p360.soap;
 
 import no.fint.p360.data.utilities.RequestUtilities;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,14 +6,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import javax.xml.namespace.QName;
 import javax.xml.ws.BindingProvider;
 
-public abstract class P360AbstractService {
+public abstract class P360AbstractSOAPService {
 
     @Autowired
     protected RequestUtilities requestUtilities;
 
     final QName serviceName;
 
-    public P360AbstractService(String namespaceURI, String localPart) {
+    public P360AbstractSOAPService(String namespaceURI, String localPart) {
         serviceName = new QName(namespaceURI, localPart);
     }
